+++ NEWS.md
@@ -9,6 +9,7 @@ taxlist 0.1.5
 * New method `update_name`, for updating information in slot `taxonNames`.
 * New method `synonyms` retrieving synonyms for indicated concepts.
 * New method `delete_name` for deleting synonyms in `taxlist` objects.
+* New method `basionym` for handling basionyms.
 
 ### Improvements